Notify the player the current game session length
Due to technical limitations of the modding API, it is not trivial to
determine the session length of the player hosting a multiplayer server
or playing in single player mode.

Players in such categories should reset the timer manually by issuing:
    /wellbeing reset-timer
at the start of each session.

For the remaining players the timer is reset automatically when the
player joins the game.